Update non-Angular packages
The most notable change here is the upgrade to version 9 of the Firebase Web SDK.
We update the imports to the v9 compat library to keep changes small for now,
but we'll want to refactor to the modular v9 style in the future.

See https://firebase.google.com/docs/web/modular-upgrade for more details.

Other changes are just small import or RxJS operator fixes.