Migrate from Google API Client Library to Identity Services library (#14)
We use the gapi.auth2 module to authenticate with Google to create a
report in Sheets. The module was deprecated, so we migrate to the newer
library in this PR. For this change, we:

- Install the `@types/google.accounts` package for the relevant typings
- Replace the library in `index.html`
- Rewrite `getGoogleSheetsAuth()` to use the new library
- Move `getGoogleSheetsAuth()` to `OauthApiService` and delete
`GapiService` to clean things up a bit
- Remove `@types/gapi` and `@types/gapi.auth2` because they are no
longer needed
- Remove references to `redirectUrls` we were not using in
`createOAuthClient()` in `serving.ts`
- Update documentation

See
https://developers.google.com/identity/oauth2/web/guides/migration-to-gis
for more details.